requirements.txt
absl-py==0.9.0
advertorch==0.2.0
boltons==20.1.0
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
google-auth==1.17.2
google-auth-oauthlib==0.4.1
grpcio==1.29.0
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.1
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.1
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
networkx==2.4
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
Pillow==7.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.4.1
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tifffile==2020.6.3
torch==1.5.0
torchvision==0.6.0a0+82fd1c8
tqdm==4.46.1
urllib3==1.25.9
Werkzeug==1.0.1
zipp==3.1.0
